Rename files in systemc and verilog
