+++ systemc/RSATwoPowerMod.cpp
