+++ verilog/RSAMontgomery.sv
@@ -1,6 +1,5 @@
-module two_power_mod #(
-	parameter MOD_WIDTH = 256,
-	parameter POWER_WIDTH = 32
+module RSAMontgomery #(
+	parameter MOD_WIDTH = 256
 ) (
 	// input
 	input clk,
@@ -9,8 +8,9 @@ module two_power_mod #(
 	// input data
 	input i_valid,
 	output i_ready,
+	input [MOD_WIDTH-1:0] i_a,
+	input [MOD_WIDTH-1:0] i_b,
 	input [MOD_WIDTH-1:0] i_modulus,
-	input [POWER_WIDTH-1:0] i_power,
 
 	// output data
 	output o_valid,
@@ -18,21 +18,25 @@ module two_power_mod #(
 	output [MOD_WIDTH-1:0] o_out
 );
 
-logic [MOD_WIDTH:0] data_modulus;
-logic [MOD_WIDTH:0] data_round_result;
-logic [POWER_WIDTH-1:0] data_power;
-logic [POWER_WIDTH-1:0] round_counter;
-logic [1:0] state, state_next;
+logic [MOD_WIDTH + 2 - 1:0] data_a;
+logic [MOD_WIDTH + 2 - 1:0] data_b;
+logic [MOD_WIDTH + 2 - 1:0] data_modulus;
+logic [MOD_WIDTH + 2 - 1:0] round_result;
+logic [MOD_WIDTH + 2 - 1:0] round_result_next;
 
-enum logic [1:0] {
+logic [$clog2(MOD_WIDTH+1)-1:0] round_counter;
+
+typedef enum logic [1:0] {
   STATE_IDLE  = 0,
   STATE_CALCULATE = 1,
   STATE_WAITDONE = 2
 } State_t;
 
+State_t state, state_next;
+
 assign i_ready = state == STATE_IDLE;
 assign o_valid = state == STATE_WAITDONE;
-assign o_out = data_round_result[MOD_WIDTH-1:0];
+assign o_out = round_result[MOD_WIDTH-1:0];
 
 // update state logic
 always_ff @(posedge clk or negedge rst) begin
@@ -53,7 +57,7 @@ always_comb begin
     end
   end
   STATE_CALCULATE: begin
-    if (round_counter == data_power - 1) begin
+    if (round_counter == MOD_WIDTH-1) begin
       state_next = STATE_WAITDONE;
     end
 	end
@@ -71,17 +75,20 @@ end
 // read input data
 always_ff @( posedge clk or negedge rst ) begin
   if (!rst) begin
-	data_power <= 0;
-	data_modulus <= 0;
+		data_a <= 0;
+		data_b <= 0;
+		data_modulus <= 0;
   end
   else begin
-	if (i_ready && i_valid) begin
-		data_power <= i_power;
-		data_modulus <= {1'b0, i_modulus};
-	end else begin
-		data_power <= data_power;
-		data_modulus <= data_modulus;
-	end
+		if (i_ready && i_valid) begin
+			data_a <= {2'b0, i_a};
+			data_b <= {2'b0, i_b};
+			data_modulus <= {2'b0, i_modulus};
+		end else begin
+			data_a <= data_a;
+			data_b <= data_b;
+			data_modulus <= data_modulus;
+		end
   end
 end
 
@@ -100,26 +107,37 @@ always_ff @(posedge clk or negedge rst) begin
   end
 end
 
-// data_round_result
+// round_result
 always_ff @(posedge clk or negedge rst) begin
   if (!rst) begin
-    data_round_result <= 1;
+    round_result <= 0;
   end
   else begin
-	case (state)
-		STATE_IDLE:
-		data_round_result <= 1;
-		STATE_CALCULATE:
-			data_round_result <= ((data_round_result << 1) > data_modulus) ?
-			(data_round_result << 1) - data_modulus :
-			data_round_result << 1;
-		STATE_WAITDONE:
-			data_round_result <= data_round_result;
-		default: begin
-			data_round_result <= 1;
-		end
-	endcase
+		round_result <= round_result_next;
+	end
+end
+
+always_comb begin
+  case (state)
+  STATE_IDLE: begin
+		round_result_next = 0;
   end
+  STATE_CALCULATE: begin
+		if (data_a[round_counter]) begin
+			round_result_next = round_result + data_b;
+		end
+		if (round_result_next[0]) begin
+			round_result_next += data_modulus;
+		end
+		round_result_next >>= 1;
+	end
+	STATE_WAITDONE: begin
+		round_result_next = round_result;
+	end
+	default: begin
+		round_result_next = 0;
+	end
+  endcase
 end
 
 endmodule
